Email has been set up, but account problems
-So the intergration has been perfect for a while but it seems the
account is suspended. This is cause we had our keys in the
appsettings.json. I contacted them to reactivate our account and I hope
it will be a fast process. I left all the intergration there and now I
am going to see if batch email sending can be done.